requirements.txt
# --- Remote Storage & Environment ---
boto3~=1.38
python-dotenv~=1.0

# --- Data Ingestion & Hashing ---
requests~=2.33
xxhash~=3.5

# --- Configuration & Metadata ---
PyYAML~=6.0
pydantic~=2.9

# --- Data Processing (Upcoming) ---
pandas~=2.2
